requirements.txt
joblib==0.17.0
numpy==1.19.2
scipy==1.5.2
pandas==1.1.2
scikit-learn==0.23.2
matplotlib==3.2.2
seaborn==0.11.0
Django==3.1.2
gunicorn==20.0.4
django-heroku==0.3.1
django-crispy-forms==1.9.2
